ml-model/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
tensorflow
pillow
